# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API Documentation",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:8080/api/v0",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/user": {
      "get": {
        "summary": "Obtenir la liste des utilisateurs",
        "description": "",
        "tags": [
          "user"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Ajouter un utilisateur",
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "John"
                },
                "password": {
                  "type": "string",
                  "example": "1234"
                },
                "email": {
                  "type": "string",
                  "example": "email@example.com"
                },
                "role": {
                  "type": "string",
                  "example": "user"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "summary": "Mettre à jour un utilisateur",
        "description": "",
        "tags": [
          "user"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/{username}/friend": {
      "get": {
        "summary": "Obtenir la liste des amis d'un user",
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "summary": "Ajouter un ami à la liste d'un user",
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
                "type": "object",
                "properties": {
                  "friendUsername": {
                    "type": "string",
                    "example": "username1"
                  }
                }
              }
            }

        ],
        "responses": {
          "201": {
            "description": "Added to friends"
          },
          "400": {
            "description": "Bad Request"
          }

        }
      }
    },
    "/user/{username}": {
      "get": {
        "summary": "Obtenir un user à partir d'un username",
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "summary": "Supprime un utilisteur à partir de son username",
        "description": "",
        "tags": [
          "user"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/ratings/{username}/": {
      "get": {
        "summary": "Obtenir les notes d'un utilisateur",
        "description": "",
        "tags": [
          "ratings"
        ],
        "parameters": [
          {
            "name": "login",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/ratings": {
      "get": {
        "summary": "Obtenir toutes les notes",
        "description": "",
        "tags": [
          "ratings"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Ajouter une note",
        "description": "",
        "tags": [
          "ratings"
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "grade": {
                  "type": "integer",
                  "example": "5",
                  "minimum": 0,
                  "maximum": 5
                },
                "review": {
                  "type": "string",
                  "example": "Super musique !"
                },
                "typeOfContent": {
                  "type": "string",
                  "example": "track"
                },
                "contentId": {
                  "type": "string",
                  "example": "5e7f3e3e3e3e3e3e3e3e3e3e"
                },
                "username": {
                  "type": "string",
                  "example": "John"
                }
              }
            }
          }
        ]
      }
    }
  }
}